Extract TensorFlowInferenceModelBase class as a base for TensorFlow models
