+++ tensorflow/src/main/kotlin/org/jetbrains/kotlinx/dl/api/inference/loaders/TFModelType.kt
@@ -1,12 +1,11 @@
 /*
- * Copyright 2022 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
+ * Copyright 2022-2023 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
 package org.jetbrains.kotlinx.dl.api.inference.loaders
 
-import org.jetbrains.kotlinx.dl.api.inference.InferenceModel
-import org.jetbrains.kotlinx.dl.api.inference.TensorFlowInferenceModel
+import org.jetbrains.kotlinx.dl.api.inference.TensorFlowInferenceModelBase
 import java.io.File
 
 /**
@@ -15,7 +14,7 @@ import java.io.File
  * @see TFModels
  * @see TFModelHub
  */
-public interface TFModelType<T : TensorFlowInferenceModel, U> : ModelType<T, U> {
+public interface TFModelType<T : TensorFlowInferenceModelBase, U> : ModelType<T, U> {
     /**
      * Loads model configuration from the provided [jsonFile].
      */
